% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{cbmod}
\alias{cbmod}
\title{Fitting models for Secondary Attack Rate with Chain Binomial response}
\usage{
cbmod(
  y,
  s0,
  x = NULL,
  i0 = 1,
  generations = Inf,
  link = "identity",
  optim_method = "BFGS"
)
}
\arguments{
\item{y}{numeric, the number of infected cases.}

\item{s0}{numeric, the number of initial susceptibles.}

\item{x}{matrix of predictors (design matrix).}

\item{i0}{numeric, number of initial infected. Default is 1.}

\item{generations}{numeric.}

\item{link}{Link function. Default is 'identity'.}

\item{optim_method}{Optimization method used by optim.}
}
\value{
A list of class `cbmod`.
}
\description{
Fitting models for Secondary Attack Rate with Chain Binomial response
}
\details{
The following link functions are available: `identity`, `log`, `logit`, and `cloglog`.
}
\examples{
set.seed(234)
mydata <- data.frame(infected = rchainbinom(n = 15, s0 = 5, sar = 0.2, i0 = 1, generations = Inf),
                     s0 = 5, i0 = 1, generations = Inf)
xmat <- model.matrix(~ 1, data = mydata)
res <- cbmod(y = mydata$infected, s0 = mydata$s0, x = xmat, i0 = mydata$i0, generations = mydata$generations)
summary(res)


}
