% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{coef.cbmod}
\alias{coef.cbmod}
\title{Extract Model Coefficient for cbmod Fits}
\usage{
\method{coef}{cbmod}(object, ...)
}
\arguments{
\item{object}{a cbmod object.}

\item{...}{other arguments. Ignored.}
}
\value{
Coefficients extracted from the cbmod object.
}
\description{
Extract Model Coefficient for cbmod Fits
}
