% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{confint.cbmod}
\alias{confint.cbmod}
\title{Confidence intervals for cbmod Object.}
\usage{
\method{confint}{cbmod}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{a cbmod object.}

\item{parm}{Character or number of which coefficient ot compute confidence
intervals for. By default intervals are computed for all coefficients.}

\item{level}{Default is 0.95, for 95\% confidence intervals.}

\item{...}{other arguments. Ignored.}
}
\value{
A two-column matrix with the lower and upper end of the confidence intervals.
}
\description{
Confidence intervals for cbmod Object.
}
