% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_sar.R
\name{confint.sar}
\alias{confint.sar}
\title{Confidence intervals for sar Object.}
\usage{
\method{confint}{sar}(object, parm = NULL, level = 0.95, method = "chisq", ...)
}
\arguments{
\item{object}{a cbmod object.}

\item{parm}{Character or number of which coefficient ot compute confidence
intervals for. By default intervals are computed for all coefficients.}

\item{level}{Default is 0.95, for 95\% confidence intervals.}

\item{method}{Either 'chisq'(default) or 'normal'.}

\item{...}{other arguments. Ignored.}
}
\value{
A numeric of length 2 with the lower and upper end of the confidence interval.
}
\description{
Confidence intervals for sar Object.
}
