% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperchainbinom.R
\name{confint.sar2}
\alias{confint.sar2}
\title{Confidence intervals for sar2 Object.}
\usage{
\method{confint}{sar2}(object, method = "chisq", level = 0.95)
}
\arguments{
\item{object}{a sar2 object.}

\item{method}{Either 'chisq'(default) or 'normal'.}

\item{level}{Default is 0.95, for 95\% confidence intervals.}

\item{parm}{Character or number of which coefficient ot compute confidence
intervals for. By default intervals are computed for all coefficients.}

\item{...}{other arguments. Ignored.}
}
\value{
A numeric of length 2 with the lower and upper end of the confidence interval.
}
\description{
Confidence intervals for sar2 Object.
}
