% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperchainbinom.R
\name{dcbhyper}
\alias{dcbhyper}
\alias{ecbhyper}
\title{The Hyper Chain Binomial distribution}
\usage{
dcbhyper(x, s0, prob, s0_obs, i0 = 1)

ecbhyper(s0, prob, s0_obs, i0 = 1)
}
\arguments{
\item{x}{numeric vector of the number of infected.}

\item{s0}{the number of initial susceptibles.}

\item{s0_obs}{The number of observed individuals.}

\item{i0}{the number of primary cases.}

\item{sar}{the secondary attack rate, or the per person risk of infection by an infected.}
}
\description{
Probability mass function, expected value, and random generation, for the hyper chain binomial distribution,
with parameters s0, sar, i0, and s0_obs for the number of observed infected
cases in a population of size s0 with s0_obs observed individuals.
}
