% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chainbinom.R, R/chainbinom2.R
\name{dchainbinom}
\alias{dchainbinom}
\alias{rchainbinom}
\alias{pchainbinom}
\alias{echainbinom}
\alias{varchainbinom}
\alias{dchainbinom2}
\title{The Chain Binomial distribution}
\usage{
dchainbinom(
  x,
  s0,
  prob,
  i0 = 1,
  generations = Inf,
  sar = lifecycle::deprecated()
)

rchainbinom(
  n,
  s0,
  prob,
  i0 = 1,
  generations = Inf,
  sar = lifecycle::deprecated()
)

pchainbinom(
  q,
  s0,
  prob,
  i0 = 1,
  generations = Inf,
  sar = lifecycle::deprecated()
)

echainbinom(s0, prob, i0 = 1, generations = Inf, sar = lifecycle::deprecated())

varchainbinom(
  s0,
  prob,
  i0 = 1,
  generations = Inf,
  sar = lifecycle::deprecated()
)

dchainbinom2(x, s0, prob, cpi)
}
\arguments{
\item{x}{Numeric vector of the number of infected.}

\item{s0}{The number of initial susceptibles.}

\item{prob}{The transmission probability.}

\item{i0}{The number of primary cases.}

\item{generations}{The number of generations. Default is Inf, which represents the entire epidemic.}

\item{sar}{Deprecated. Please use the argument 'prob' instead.}

\item{n}{Number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{q}{Numeric vector of the number of infected.}

\item{cpi}{The community probability of infection. .}
}
\value{
\code{dchainbinom} gives the probability of x infected, given s0, i0, prob and generations. \code{echainbinom} gives the expected value,
\code{varchainbinom} gives the variance, \code{rchainbinom} generates random data.

\code{dchainbinom2} gives the probability of x infected, given s0, prob and cpi.
}
\description{
Probability mass function, cumulative, expected value, variance function, and random generation, for the chain binomial distribution
for the number of infected cases in a population of size s0. The \code{chainbinom} family has parameters s0, i0, prob and number of generations,
and corresponds to the classical Reed-Frost model. The \code{chainbinom2} family has parameters prob and cpi, and is the Longini-Koopman model.
}
\examples{
dchainbinom(x = 0:5, s0 = 5, prob = 0.2, i0 = 1, generations = Inf)
rchainbinom(n = 10, s0 = 5, prob = 0.2, i0 = 1, generations = Inf)
echainbinom(s0 = 5, prob = 0.2, i0 = 1, generations = Inf)
varchainbinom(s0 = 5, prob = 0.2, i0 = 1, generations = Inf)

}
