% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperchainbinom.R
\name{estimate_sar_cbhyper}
\alias{estimate_sar_cbhyper}
\title{Estimate Secondary Attack Rate of the Hyper Chain Binomial Model}
\usage{
estimate_sar_cbhyper(infected, s0, s0_obs, i0 = 1, se = TRUE)
}
\arguments{
\item{infected}{numeric.}

\item{s0}{numeric.}

\item{s0_obs}{numeric.}

\item{i0}{numeric.}

\item{se}{logical. If TRUE (default), the standard error is computed.}
}
\value{
A list of class \code{sar2} with the following components:
\itemize{
\item \code{sar_hat} The point estimate of the secondary attack rate.
\item \code{se} Standard error of the estimate (if se = TRUE).
\item \code{loglikelihood} the log likelihood value at the point estimate.
\item \code{data} the input data.
}
}
\description{
Given data on the number of infected after a number of generation, initial
number of susceptible, and initial number of infected, estimate the
secondary attack rate (SAR) using maximum likelihood.
}
\seealso{
\code{\link[=confint.sar2]{confint.sar2()}} for calculating confidence intervals.
}
