% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{glance.cbmod}
\alias{glance.cbmod}
\title{Glance at a cbmod object}
\usage{
\method{glance}{cbmod}(x, ...)
}
\arguments{
\item{x}{A \code{cbmod} object.}

\item{...}{Unused.}
}
\value{
Returns a tibble with the following columns:
\itemize{
\item \code{logLik} The model's log-likelihood.
\item \code{npar} Numnber of parameters in the model.
}
}
\description{
Glance at a cbmod object
}
