% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{residuals.cbmod}
\alias{residuals.cbmod}
\title{Extract Model Residuals for cbmod Fits}
\usage{
\method{residuals}{cbmod}(object, type = "response", ...)
}
\arguments{
\item{object}{a fitted object of class inheriting from "cbmod".}

\item{type}{he type of residuals which should be returned. The alternatives are: "response" (default), "far", and "pearson".}

\item{...}{additional arguments.}
}
\value{
a vector of residuals.
}
\description{
Extract Model Residuals for cbmod Fits
}
