% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{summary.cbmod}
\alias{summary.cbmod}
\title{Summary of cbmod Object.}
\usage{
\method{summary}{cbmod}(object, ...)
}
\arguments{
\item{object}{a cbmod object.}

\item{...}{other arguments. Ignored.}
}
\value{
Returns nothing, but displays a summary of the model fit.
}
\description{
Summary of cbmod Object.
}
