% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{vcov.cbmod}
\alias{vcov.cbmod}
\title{Variance-Covariance Matrix of cbmod Object.}
\usage{
\method{vcov}{cbmod}(object, ...)
}
\arguments{
\item{object}{a cbmod object.}

\item{...}{other arguments. Ignored.}
}
\value{
A variance-covariance matrix.
}
\description{
Variance-Covariance Matrix of cbmod Object.
}
