% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_sar.R
\name{estimate_sar}
\alias{estimate_sar}
\title{Estimate Secondary Attack Rate of the Chain Binomial Model}
\usage{
estimate_sar(infected, s0, i0 = 1, generations = Inf, se = TRUE)
}
\arguments{
\item{infected}{numeric.}

\item{s0}{numeric.}

\item{i0}{numeric.}

\item{generations}{numeric.}
}
\description{
Given data on the number of infected after a number of generation, initial
number of susceptible, and initial number of infected, estimate the
secondary attack rate (SAR) using maximum likelihood.
}
