% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{glance.cbmod}
\alias{glance.cbmod}
\title{Glance at a cbmod object}
\usage{
\method{glance}{cbmod}(x, ...)
}
\arguments{
\item{x}{A `cbmod` object.}

\item{...}{Unused.}
}
\description{
Glance at a cbmod object
}
