% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy.cbmod}
\alias{tidy.cbmod}
\title{Tidy a cbmod object}
\usage{
\method{tidy}{cbmod}(x, conf.int = FALSE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{A `lm` object.}

\item{conf.int}{Logical indicating whether or not to include a confidence interval in the tidied output. Defaults to FALSE.}

\item{conf.level}{The confidence level to use for the confidence interval if conf.int = TRUE. Defaults to 0.95, which corresponds to a
95 percent confidence interval.}

\item{...}{Unused.}
}
\description{
Tidy a cbmod object
}
